backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3
python-dotenv==1.0.0
httpx==0.26.0
